skills_dashboard/requirements.txt
Scrapy==2.4.1
matplotlib==3.3.1
itemadapter==0.2.0
Django==3.1.2
drf_yasg==1.20.0
djangorestframework==3.12.4
djongo==1.3.6
sqlparse==0.2.4
